download_prepare_clean_normalize_sedici_dataset/requirements.txt
requests
pandas
google-genai
openai
python-dotenv
pdfplumber
PyMuPDF
pytesseract
pdf2image
chardet